refactor: PositionSelect 공통 ui로 분리
