+++ src/components/user/profile/ProfileForm.tsx
@@ -18,7 +18,7 @@ import { useSetRecoilState } from 'recoil';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { isEqual } from 'lodash';
 import { snackbarState } from '@/store/CommonStateStore';
-import PositionSelect from '@/components/user/signup/PositionSelect';
+import PositionSelect from '@/components/ui/selector/PositionSelect';
 import TechStackSelect from '@/components/ui/form/TechStackSelect';
 
 function ProfileForm({ profileInfo }: { profileInfo: ProfileInfo }) {
